fix: fix all code in auth system refactor: refactor logic in auth system
