+++ lib/features/auth/models/requests/forgot_password_request.dart
@@ -1,0 +1,9 @@
+class ForgotPasswordRequest {
+  final String email;
+
+  ForgotPasswordRequest({required this.email});
+
+  Map<String, dynamic> toJson() {
+    return {'email': email};
+  }
+}
