+++ lib/features/auth/models/requests/verify_otp_request.dart
@@ -1,0 +1,10 @@
+class VerifyOtpRequest {
+  final String email;
+  final String otp;
+
+  VerifyOtpRequest({required this.email, required this.otp});
+
+  Map<String, dynamic> toJson() {
+    return {'email': email, 'otp': otp};
+  }
+}
