--- lib/features/auth/models/reset_password_response.dart
@@ -1,13 +1,0 @@
-class ResetPasswordResponse {
-  final bool success;
-  final String message;
-
-  ResetPasswordResponse({required this.success, required this.message});
-
-  factory ResetPasswordResponse.fromJson(Map<String, dynamic> json) {
-    return ResetPasswordResponse(
-      success: json['success'] ?? false,
-      message: json['message'] ?? '',
-    );
-  }
-}
