+++ lib/features/auth/models/responses/forgot_password_response.dart
@@ -1,0 +1,13 @@
+class ForgotPasswordResponse {
+  final String status;
+  final String? data;
+
+  ForgotPasswordResponse({required this.status, this.data});
+
+  factory ForgotPasswordResponse.fromJson(Map<String, dynamic> json) {
+    return ForgotPasswordResponse(
+      status: json['status'],
+      data: json['data'],
+    );
+  }
+}
