+++ lib/features/auth/models/responses/register_response.dart
@@ -1,0 +1,13 @@
+class RegisterResponse {
+  final String status;
+  final String? data;
+
+  RegisterResponse({required this.status, this.data});
+
+  factory RegisterResponse.fromJson(Map<String, dynamic> json) {
+    return RegisterResponse(
+      status: json['status'],
+      data: json['data'],
+    );
+  }
+}
