+++ lib/features/auth/models/responses/reset_password_response.dart
@@ -1,0 +1,11 @@
+class ResetPasswordResponse {
+  final String status;
+
+  ResetPasswordResponse({required this.status});
+
+  factory ResetPasswordResponse.fromJson(Map<String, dynamic> json) {
+    return ResetPasswordResponse(
+      status: json['status'],
+    );
+  }
+}
