--- lib/features/auth/models/verify_opt_request.dart
@@ -1,8 +1,0 @@
-class VerifyOtpRequest {
-  final String email;
-  final String opt;
-  VerifyOtpRequest({required this.email, required this.opt});
-  Map<String, dynamic> toJson() {
-    return {'email': email, 'opt': opt};
-  }
-}
