+++ lib/features/auth/view/reset_password_screen.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:hit_tech/core/constants/app_string.dart';
-import 'package:hit_tech/features/auth/models/requests/forgot_password_request.dart';
-import 'package:hit_tech/features/auth/models/responses/forgot_password_response.dart';
+import 'package:hit_tech/features/auth/models/requests/reset_password_request.dart';
+import 'package:hit_tech/features/auth/models/responses/reset_password_response.dart';
+import 'package:hit_tech/features/auth/view/login_screen.dart';
 import 'package:hit_tech/features/auth/view/widgets/custom_input_field.dart';
 
 import '../../../core/constants/app_assets.dart';
@@ -9,46 +10,20 @@ import '../../../core/constants/app_color.dart';
 import '../../../core/constants/app_message.dart';
 import '../service/auth_service.dart';
 import '../utils/validator_util.dart';
-import 'otp_verification_screen.dart';
 
-class ForgotPasswordScreen extends StatefulWidget {
-  const ForgotPasswordScreen({super.key});
+class ResetPasswordScreen extends StatefulWidget {
+  final String email;
+
+  const ResetPasswordScreen({super.key, required this.email});
 
   @override
-  State<ForgotPasswordScreen> createState() => _ForgotPasswordScreenState();
+  State<ResetPasswordScreen> createState() => _ResetPasswordScreenState();
 }
 
-class _ForgotPasswordScreenState extends State<ForgotPasswordScreen> {
-  final TextEditingController _emailController = TextEditingController();
+class _ResetPasswordScreenState extends State<ResetPasswordScreen> {
+  final TextEditingController _newPassword = TextEditingController();
+  final TextEditingController _reEnterNewPassword = TextEditingController();
   final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
-
-  Future<void> _handleForgot() async {
-    if (!_formKey.currentState!.validate()) return;
-    try {
-      final request = ForgotPasswordRequest(
-        email: _emailController.text.trim(),
-      );
-      final ForgotPasswordResponse response =
-          await AuthService.sendEmailToForgotPassword(request);
-
-      if (response.status == 'SUCCESS') {
-        Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (context) =>
-                OtpVerificationScreen(email: request.email, isRegister: false),
-          ),
-        );
-      } else {
-        _showSnackBar(
-          response.data ?? UserMessage.errEmailNotExist,
-          isError: true,
-        );
-      }
-    } catch (e) {
-      _showSnackBar(UserMessage.errEmailNotExist, isError: false);
-    }
-  }
 
   void _showSnackBar(String message, {bool isError = false}) {
     ScaffoldMessenger.of(context).showSnackBar(
@@ -60,10 +35,37 @@ class _ForgotPasswordScreenState extends State<ForgotPasswordScreen> {
     );
   }
 
+  Future<void> _handleResetPassword() async {
+    if (!_formKey.currentState!.validate()) return;
+    try {
+      final request = ResetPasswordRequest(
+        email: widget.email,
+        newPassword: _newPassword.text,
+        reEnterPassword: _reEnterNewPassword.text,
+      );
+      final ResetPasswordResponse response = await AuthService.resetPassword(
+        request,
+      );
+
+      if (response.status == 'SUCCESS') {
+        Navigator.push(
+          context,
+          MaterialPageRoute(builder: (context) => LoginScreen()),
+        );
+      } else {
+        _showSnackBar(
+          AppStrings.generalError,
+          isError: true,
+        );
+      }
+    } catch (e) {
+      _showSnackBar(HttpMessage.errServer, isError: true);
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     final double screenWidth = MediaQuery.of(context).size.width;
-
     return Scaffold(
       backgroundColor: Colors.white,
       body: Stack(
@@ -77,32 +79,18 @@ class _ForgotPasswordScreenState extends State<ForgotPasswordScreen> {
           Padding(
             padding: const EdgeInsets.symmetric(horizontal: 24.0, vertical: 51),
             child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                Row(
-                  children: [
-                    IconButton(
-                      onPressed: () {
-                        Navigator.pop(context);
-                      },
-                      icon: Icon(
-                        Icons.arrow_back_ios,
-                        color: AppColors.bNormal,
-                        size: 28,
-                      ),
-                    ),
-                    const SizedBox(width: 30),
-                    const Text(
-                      "Quên mật khẩu",
-                      style: TextStyle(
-                        fontSize: 32,
-                        fontWeight: FontWeight.w500,
-                        color: AppColors.dark,
-                      ),
-                    ),
-                  ],
+                const Text(
+                  "Quên mật khẩu",
+                  style: TextStyle(
+                    fontSize: 32,
+                    fontWeight: FontWeight.w500,
+                    color: AppColors.dark,
+                  ),
                 ),
+
                 const SizedBox(height: 16),
+
                 Row(
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: List.generate(3, (index) {
@@ -111,29 +99,47 @@ class _ForgotPasswordScreenState extends State<ForgotPasswordScreen> {
                       width: 32,
                       height: 6,
                       decoration: BoxDecoration(
-                        color: index == 0
-                            ? AppColors.bNormal
-                            : Colors.grey[400],
+                        color: AppColors.bNormal,
                         borderRadius: BorderRadius.circular(20),
                       ),
                     );
                   }),
                 ),
+
                 const SizedBox(height: 16),
 
-                //content
+                Text(
+                  AppStrings.resetYourPassword,
+                  style: TextStyle(color: AppColors.dark, fontSize: 16),
+                ),
+
+                const SizedBox(height: 40),
+
                 Form(
                   key: _formKey,
                   child: Center(
                     child: Column(
                       crossAxisAlignment: CrossAxisAlignment.center,
                       children: [
-                        Text(
-                          AppStrings.fillEmailToResetPassword,
-                          style: TextStyle(color: AppColors.dark, fontSize: 16),
+                        // Text Field
+                        CustomInputField(
+                          isPassword: true,
+                          hintStyle: const TextStyle(
+                            color: Colors.grey,
+                            fontSize: 16,
+                          ),
+                          width: screenWidth * 0.9,
+                          height: 64,
+                          controller: _newPassword,
+                          title: AppStrings.password,
+                          borderRadius: 12,
+                          borderColor: Colors.grey[400],
+                          focusedBorderColor: AppColors.bNormal,
+                          validator: ValidatorUtil.validatePassword,
+                          onChanged: (value) {},
                         ),
 
-                        SizedBox(height: 40),
+                        SizedBox(height: 24),
 
                         // Text Field
                         CustomInputField(
@@ -144,12 +150,12 @@ class _ForgotPasswordScreenState extends State<ForgotPasswordScreen> {
                           ),
                           width: screenWidth * 0.9,
                           height: 64,
-                          controller: _emailController,
-                          title: AppStrings.email,
+                          controller: _reEnterNewPassword,
+                          title: AppStrings.reEnterPassword,
                           borderRadius: 12,
                           borderColor: Colors.grey[400],
                           focusedBorderColor: AppColors.bNormal,
-                          validator: ValidatorUtil.validateEmail,
+                          validator: ValidatorUtil.validatePassword,
                           onChanged: (value) {},
                         ),
 
@@ -160,10 +166,11 @@ class _ForgotPasswordScreenState extends State<ForgotPasswordScreen> {
                           width: MediaQuery.of(context).size.width * 0.9,
                           height: 50,
                           child: ElevatedButton(
-                            onPressed: _handleForgot,
+                            onPressed: _handleResetPassword,
                             style: ElevatedButton.styleFrom(
                               backgroundColor:
-                                  (_emailController.text.isNotEmpty)
+                                  (_reEnterNewPassword.text.isNotEmpty &&
+                                      _newPassword.text.isNotEmpty)
                                   ? AppColors.bNormal
                                   : AppColors.lighter,
                               disabledBackgroundColor: Colors.grey[400],
@@ -176,7 +183,9 @@ class _ForgotPasswordScreenState extends State<ForgotPasswordScreen> {
                               style: TextStyle(
                                 fontSize: 20,
                                 fontWeight: FontWeight.bold,
-                                color: (_emailController.text.isNotEmpty)
+                                color:
+                                    (_reEnterNewPassword.text.isNotEmpty &&
+                                        _newPassword.text.isNotEmpty)
                                     ? AppColors.wWhite
                                     : AppColors.moreLighter,
                               ),
