--- lib/features/auth/view/widgets/opt_verification_screen.dart
@@ -1,393 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:hit_tech/features/auth/models/verify_opt_request.dart';
-import 'dart:async';
-
-import 'package:hit_tech/features/auth/models/repositories/auth_repository.dart';
-
-class OtpVerificationScreen extends StatefulWidget {
-  final String email;
-
-  const OtpVerificationScreen({Key? key, required this.email})
-    : super(key: key);
-
-  @override
-  State<OtpVerificationScreen> createState() => _OtpVerificationScreenState();
-}
-
-class _OtpVerificationScreenState extends State<OtpVerificationScreen> {
-  final List<TextEditingController> _controllers = List.generate(
-    6,
-    (index) => TextEditingController(),
-  );
-  final List<FocusNode> _focusNodes = List.generate(6, (index) => FocusNode());
-
-  bool _isLoading = false;
-  bool _isResending = false;
-  int _remainingTime = 60;
-  Timer? _timer;
-
-  @override
-  void initState() {
-    super.initState();
-    _startTimer();
-  }
-
-  @override
-  void dispose() {
-    _timer?.cancel();
-    for (var controller in _controllers) {
-      controller.dispose();
-    }
-    for (var focusNode in _focusNodes) {
-      focusNode.dispose();
-    }
-    super.dispose();
-  }
-
-  void _startTimer() {
-    _remainingTime = 60;
-    _timer = Timer.periodic(const Duration(seconds: 1), (timer) {
-      setState(() {
-        if (_remainingTime > 0) {
-          _remainingTime--;
-        } else {
-          _timer?.cancel();
-        }
-      });
-    });
-  }
-
-  String get _formattedTime {
-    int minutes = _remainingTime ~/ 60;
-    int seconds = _remainingTime % 60;
-    return '${minutes.toString().padLeft(2, '0')}:${seconds.toString().padLeft(2, '0')}';
-  }
-
-  void _onOtpDigitChanged(String value, int index) {
-    if (value.length == 1) {
-      // Move to next field
-      if (index < 5) {
-        _focusNodes[index + 1].requestFocus();
-      } else {
-        // Auto verify when last digit is entered
-        _verifyOtp();
-      }
-    } else if (value.isEmpty && index > 0) {
-      // Move to previous field when deleting
-      _focusNodes[index - 1].requestFocus();
-    }
-  }
-
-  void _onOtpDigitBackspace(int index) {
-    if (_controllers[index].text.isEmpty && index > 0) {
-      _focusNodes[index - 1].requestFocus();
-    }
-  }
-
-  String _getOtpCode() {
-    return _controllers.map((controller) => controller.text).join();
-  }
-
-  bool _isOtpComplete() {
-    return _getOtpCode().length == 6;
-  }
-
-  Future<void> _verifyOtp() async {
-    if (!_isOtpComplete()) {
-      _showMessage('Vui lòng nhập đầy đủ mã OTP', isError: true);
-      return;
-    }
-
-    setState(() {
-      _isLoading = true;
-    });
-
-    try {
-      final authRepository = AuthRepository();
-      final request = VerifyOtpRequest(email: widget.email, opt: _getOtpCode());
-
-      final response = await authRepository.verifyOtp(request);
-
-      if (response.success) {
-        _showMessage('Xác thực thành công!');
-        // Navigate to home screen or next screen
-        Navigator.pushReplacementNamed(context, '/home');
-      } else {
-        _showMessage(response.message ?? 'Xác thực thất bại', isError: true);
-        _clearOtp();
-      }
-    } catch (e) {
-      _showMessage('Có lỗi xảy ra. Vui lòng thử lại.', isError: true);
-    } finally {
-      setState(() {
-        _isLoading = false;
-      });
-    }
-  }
-
-  Future<void> _resendOtp() async {
-    if (_remainingTime > 0) return;
-
-    setState(() {
-      _isResending = true;
-    });
-
-    try {
-      // Call resend OTP API here
-      await Future.delayed(const Duration(seconds: 2)); // Simulate API call
-
-      _showMessage('Mã OTP mới đã được gửi');
-      _clearOtp();
-      _startTimer();
-    } catch (e) {
-      _showMessage(
-        'Không thể gửi lại mã OTP. Vui lòng thử lại.',
-        isError: true,
-      );
-    } finally {
-      setState(() {
-        _isResending = false;
-      });
-    }
-  }
-
-  void _clearOtp() {
-    for (var controller in _controllers) {
-      controller.clear();
-    }
-    _focusNodes[0].requestFocus();
-  }
-
-  void _showMessage(String message, {bool isError = false}) {
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Text(message),
-        backgroundColor: isError ? Colors.red : Colors.green,
-        duration: const Duration(seconds: 3),
-      ),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        backgroundColor: Colors.transparent,
-        elevation: 0,
-        leading: IconButton(
-          icon: const Icon(Icons.arrow_back, color: Colors.black),
-          onPressed: () => Navigator.pop(context),
-        ),
-      ),
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.all(24.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              const SizedBox(height: 40),
-
-              // Icon
-              Container(
-                width: 80,
-                height: 80,
-                decoration: BoxDecoration(
-                  color: Colors.blue.shade50,
-                  borderRadius: BorderRadius.circular(40),
-                ),
-                child: Icon(
-                  Icons.email_outlined,
-                  size: 40,
-                  color: Colors.blue.shade600,
-                ),
-              ),
-
-              const SizedBox(height: 32),
-
-              // Title
-              const Text(
-                'Xác thực OTP',
-                style: TextStyle(
-                  fontSize: 28,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.black,
-                ),
-              ),
-
-              const SizedBox(height: 16),
-
-              // Description
-              Text(
-                'Chúng tôi đã gửi mã xác thực 6 chữ số đến\n${widget.email}',
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                  fontSize: 16,
-                  color: Colors.grey.shade600,
-                  height: 1.5,
-                ),
-              ),
-
-              const SizedBox(height: 48),
-
-              // OTP Input Fields
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: List.generate(6, (index) {
-                  return Container(
-                    width: 48,
-                    height: 56,
-                    decoration: BoxDecoration(
-                      border: Border.all(
-                        color: _controllers[index].text.isNotEmpty
-                            ? Colors.blue.shade600
-                            : Colors.grey.shade300,
-                        width: 2,
-                      ),
-                      borderRadius: BorderRadius.circular(12),
-                    ),
-                    child:
-                        TextField(
-                          controller: _controllers[index],
-                          focusNode: _focusNodes[index],
-                          textAlign: TextAlign.center,
-                          keyboardType: TextInputType.number,
-                          maxLength: 1,
-                          style: const TextStyle(
-                            fontSize: 20,
-                            fontWeight: FontWeight.bold,
-                          ),
-                          decoration: const InputDecoration(
-                            border: InputBorder.none,
-                            counterText: '',
-                          ),
-                          inputFormatters: [
-                            FilteringTextInputFormatter.digitsOnly,
-                          ],
-                          onChanged: (value) =>
-                              _onOtpDigitChanged(value, index),
-                          onTap: () {
-                            _controllers[index].selection =
-                                TextSelection.fromPosition(
-                                  TextPosition(
-                                    offset: _controllers[index].text.length,
-                                  ),
-                                );
-                          },
-                        ).onKeyEvent(
-                          onKey: (node, event) {
-                            if (event is KeyUpEvent &&
-                                event.logicalKey ==
-                                    LogicalKeyboardKey.backspace) {
-                              _onOtpDigitBackspace(index);
-                            }
-                            return KeyEventResult.ignored;
-                          },
-                        ),
-                  );
-                }),
-              ),
-
-              const SizedBox(height: 32),
-
-              // Timer and Resend
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Text(
-                    'Gửi lại mã sau ',
-                    style: TextStyle(fontSize: 16, color: Colors.grey.shade600),
-                  ),
-                  Text(
-                    _formattedTime,
-                    style: TextStyle(
-                      fontSize: 16,
-                      fontWeight: FontWeight.bold,
-                      color: Colors.blue.shade600,
-                    ),
-                  ),
-                ],
-              ),
-
-              const SizedBox(height: 16),
-
-              // Resend Button
-              TextButton(
-                onPressed: _remainingTime == 0 && !_isResending
-                    ? _resendOtp
-                    : null,
-                child: _isResending
-                    ? const SizedBox(
-                        width: 20,
-                        height: 20,
-                        child: CircularProgressIndicator(strokeWidth: 2),
-                      )
-                    : Text(
-                        'Gửi lại mã OTP',
-                        style: TextStyle(
-                          fontSize: 16,
-                          fontWeight: FontWeight.w600,
-                          color: _remainingTime == 0
-                              ? Colors.blue.shade600
-                              : Colors.grey.shade400,
-                        ),
-                      ),
-              ),
-
-              const Spacer(),
-
-              // Verify Button
-              SizedBox(
-                width: double.infinity,
-                height: 56,
-                child: ElevatedButton(
-                  onPressed: _isOtpComplete() && !_isLoading
-                      ? _verifyOtp
-                      : null,
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: Colors.blue.shade600,
-                    foregroundColor: Colors.white,
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(12),
-                    ),
-                    elevation: 0,
-                  ),
-                  child: _isLoading
-                      ? const SizedBox(
-                          width: 24,
-                          height: 24,
-                          child: CircularProgressIndicator(
-                            strokeWidth: 2,
-                            valueColor: AlwaysStoppedAnimation<Color>(
-                              Colors.white,
-                            ),
-                          ),
-                        )
-                      : const Text(
-                          'Xác thực',
-                          style: TextStyle(
-                            fontSize: 18,
-                            fontWeight: FontWeight.w600,
-                          ),
-                        ),
-                ),
-              ),
-
-              const SizedBox(height: 32),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-// Extension for onKeyEvent
-extension TextFieldExtension on TextField {
-  Widget onKeyEvent({
-    required KeyEventResult Function(FocusNode, KeyEvent) onKey,
-  }) {
-    return Focus(onKeyEvent: onKey, child: this);
-  }
-}
