+++ lib/main.dart
@@ -1,15 +1,16 @@
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:hit_tech/features/auth/models/repositories/auth_repository.dart';
-import 'package:hit_tech/features/auth/cubit/login-register/blocs/register_cubit.dart';
 import 'package:hit_tech/features/auth/view/forgot_password_screen.dart';
+import 'package:hit_tech/features/auth/view/reset_password_screen.dart';
 
 import 'package:hit_tech/features/auth/view/splash_page.dart';
-import 'package:hit_tech/features/auth/view/login_page.dart';
+import 'package:hit_tech/features/auth/view/login_screen.dart';
 import 'package:hit_tech/features/auth/view/register_screen.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:hit_tech/features/auth/view/otp_verification_screen.dart';
 import 'package:hit_tech/features/health_infor/view/health_info_page.dart';
+import 'package:hit_tech/features/health_infor/view/widgets/gender_selection_widget.dart';
 import 'package:hit_tech/features/home/view/home_screen.dart';
 import 'package:hit_tech/features/training_flow/view/widget/training_equipment_selection_widget.dart';
 import 'package:hit_tech/features/training_flow/view/widget/training_frequency_selection_widget.dart';
@@ -40,10 +41,6 @@ class MyApp extends StatelessWidget {
       builder: (context, child) {
         return MultiBlocProvider(
           providers: [
-            BlocProvider<AuthBloc>(
-              create: (context) => AuthBloc(authRepository: AuthRepository()),
-            ),
-
             BlocProvider<HealthInfoBloc>(
               create: (context) => HealthInfoBloc(HealthInforRepo(Dio())),
               child: HealthInfoPage(),
@@ -58,16 +55,16 @@ class MyApp extends StatelessWidget {
               fontFamily: 'Roboto',
               textTheme: Typography.englishLike2018.apply(fontSizeFactor: 1.sp),
             ),
-            home: TrainingGoalSelectionWidget(),
+            home: LoginScreen(),
             routes: {
-              '/login': (context) => const LoginScreen(),
+              // '/login': (context) => const LoginScreen(),
               '/register': (context) => const RegisterScreen(),
               '/home': (context) => HomeScreen(),
-              '/forgot-password': (context) => const ForgotPasswordScreen(),
+              // '/forgot-password': (context) => const ForgotPasswordScreen(),
               '/splash': (context) => const SplashScreen(),
               '/homeRoot': (context) => HomeRoot(),
-              '/trainingGoalSelection': (context) =>
-                  TrainingGoalSelectionWidget(),
+              // '/trainingGoalSelection': (context) =>
+              //     TrainingGoalSelectionWidget(),
               // '/trainingTypeSelection': (context) =>
               //     TrainingTypeSelectionWidget(),
               // '/trainingDurationSelection': (context) =>
